pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roverdatakit"
version = "0.1.0"
description = "Production-grade data pipeline for autonomous rover dataset collection and processing"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "PyQt6>=6.7.0",
    "PyQt6-Qt6>=6.7.0",
    "PyQt6-WebEngine>=6.7.0",
    "PyQtGraph>=0.13.7",
    "opencv-python>=4.10.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-image>=0.23.0",
    "pandas>=2.2.0",
    "py-gpmf-parser>=0.1.1",
    "pydantic>=2.7.0",
    "loguru>=0.7.2",
    "toml>=0.10.2",
    "Pillow>=10.3.0",
    "folium>=0.17.0",
]

[project.optional-dependencies]
ai = [
    "ultralytics>=8.2.0",
    "transformers>=4.42.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "albumentations>=1.4.8",
    "imagehash>=4.3.2",
    "evo>=1.29.0",
]
web = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "celery>=5.4.0",
    "redis>=5.0.4",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "Jinja2>=3.1.4",
]

[project.scripts]
roverdatakit = "desktop.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "desktop*", "web*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
